Creating unique keys for identical identifiers.
For example, the sentence 'take the cup off the table and bring it to
the shelf' has six different parses, four labeled with 'smain', and two
labeled with 'smain/.r'. Previously, since the labels were used for the
dictionary keys, only two parses were returned (and the others
overwritten).
Now, the parses are labeled as smain, smain/0, smain/.r, smain/.r/0,
etc. to provide unique keys.